# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BookReviver",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "description": "List the projects of the signed-in account, most recently updated first.",
        "operationId": "list_projects_api_v1_projects_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProjectSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "client-error-type",
                  "title": "User facing error message.",
                  "status": 400,
                  "detail": "Additional error context."
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "server-error-type",
                  "title": "User facing error message.",
                  "status": 500,
                  "detail": "Additional error context."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "description": "Create a project from the description of its book, and point to it in the ``Location`` header.",
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "client-error-type",
                  "title": "User facing error message.",
                  "status": 400,
                  "detail": "Additional error context."
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "server-error-type",
                  "title": "User facing error message.",
                  "status": 500,
                  "detail": "Additional error context."
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Project",
        "description": "Return one project of the signed-in account.",
        "operationId": "project_api_v1_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Identifier of the project",
              "title": "Project Id"
            },
            "description": "Identifier of the project"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "client-error-type",
                  "title": "User facing error message.",
                  "status": 400,
                  "detail": "Additional error context."
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "server-error-type",
                  "title": "User facing error message.",
                  "status": 500,
                  "detail": "Additional error context."
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update Project",
        "description": "Merge the body into the project, as JSON Merge Patch (RFC 7396) defines.\n\nA field left out keeps its value and a field sent as null is cleared to the value a new project has for it; the\ntitle cannot be cleared, and the cover must be a page of the project.",
        "operationId": "update_project_api_v1_projects__project_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Identifier of the project",
              "title": "Project Id"
            },
            "description": "Identifier of the project"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "client-error-type",
                  "title": "User facing error message.",
                  "status": 400,
                  "detail": "Additional error context."
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "server-error-type",
                  "title": "User facing error message.",
                  "status": 500,
                  "detail": "Additional error context."
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project",
        "description": "Delete a project with its pages, jobs and files.",
        "operationId": "delete_project_api_v1_projects__project_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Identifier of the project",
              "title": "Project Id"
            },
            "description": "Identifier of the project"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "client-error-type",
                  "title": "User facing error message.",
                  "status": 400,
                  "detail": "Additional error context."
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "server-error-type",
                  "title": "User facing error message.",
                  "status": 500,
                  "detail": "Additional error context."
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Read Job",
        "description": "Return the state and progress of a job.",
        "operationId": "read_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Identifier of the job",
              "title": "Job Id"
            },
            "description": "Identifier of the job"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "client-error-type",
                  "title": "User facing error message.",
                  "status": 400,
                  "detail": "Additional error context."
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "server-error-type",
                  "title": "User facing error message.",
                  "status": 500,
                  "detail": "Additional error context."
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel Job",
        "description": "Cancel a queued or running job and return it; a running job stops before its next step.",
        "operationId": "cancel_job_api_v1_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Identifier of the job",
              "title": "Job Id"
            },
            "description": "Identifier of the job"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "client-error-type",
                  "title": "User facing error message.",
                  "status": 400,
                  "detail": "Additional error context."
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "server-error-type",
                  "title": "User facing error message.",
                  "status": 500,
                  "detail": "Additional error context."
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/events": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Project-Events",
        "description": "Stream the project's events as they happen: jobs, imported sources, ready scans, pages, versions, description.",
        "operationId": "project_events_api_v1_projects__project_id__events_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Identifier of the project",
              "title": "Project Id"
            },
            "description": "Identifier of the project"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "client-error-type",
                  "title": "User facing error message.",
                  "status": 400,
                  "detail": "Additional error context."
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "server-error-type",
                  "title": "User facing error message.",
                  "status": 500,
                  "detail": "Additional error context."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Problem title"
          },
          "type": {
            "type": "string",
            "title": "Problem type"
          },
          "status": {
            "type": "integer",
            "title": "Status code"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "errors"
        ],
        "title": "RequestValidationError"
      },
...
      "Page_ProjectSchema_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectSchema"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[ProjectSchema]"
      },
...
      "ProjectCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title"
          },
          "subtitle": {
            "type": "string",
            "maxLength": 300,
            "title": "Subtitle",
            "default": ""
          },
          "parallel_titles": {
            "items": {
              "type": "string",
              "maxLength": 300,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 10,
            "title": "Parallel Titles"
          },
          "original_title": {
            "type": "string",
            "maxLength": 300,
            "title": "Original Title",
            "default": ""
          },
          "contributors": {
            "items": {
              "$ref": "#/components/schemas/ContributorSchema"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Contributors"
          },
          "publisher": {
            "type": "string",
            "maxLength": 300,
            "title": "Publisher",
            "default": ""
          },
          "printer": {
            "type": "string",
            "maxLength": 300,
            "title": "Printer",
            "default": ""
          },
          "publication_place": {
            "type": "string",
            "maxLength": 300,
            "title": "Publication Place",
            "default": ""
          },
          "publication_year": {
            "type": "string",
            "maxLength": 300,
            "title": "Publication Year",
            "default": ""
          },
          "edition": {
            "type": "string",
            "maxLength": 300,
            "title": "Edition",
            "default": ""
          },
          "censorship": {
            "type": "string",
            "maxLength": 300,
            "title": "Censorship",
            "default": ""
          },
          "series": {
            "type": "string",
            "maxLength": 300,
            "title": "Series",
            "default": ""
          },
          "series_number": {
            "type": "string",
            "maxLength": 300,
            "title": "Series Number",
            "default": ""
          },
          "volume": {
            "type": "string",
            "maxLength": 300,
            "title": "Volume",
            "default": ""
          },
          "languages": {
            "items": {
              "type": "string",
              "pattern": "^[a-z]{3}$"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Languages"
          },
          "orthography": {
            "$ref": "#/components/schemas/Orthography",
            "default": "unknown"
          },
          "script": {
            "$ref": "#/components/schemas/Script",
            "default": "unknown"
          },
          "printed_pagination": {
            "type": "string",
            "maxLength": 300,
            "title": "Printed Pagination",
            "default": ""
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 200.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "illustrations": {
            "type": "string",
            "maxLength": 300,
            "title": "Illustrations",
            "default": ""
          },
          "binding": {
            "type": "string",
            "maxLength": 300,
            "title": "Binding",
            "default": ""
          },
          "identifiers": {
            "items": {
              "$ref": "#/components/schemas/IdentifierSchema"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Identifiers"
          },
          "subjects": {
            "items": {
              "type": "string",
              "maxLength": 300,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 50,
            "title": "Subjects"
          },
          "rights": {
            "$ref": "#/components/schemas/RightsStatus",
            "default": "unknown"
          },
          "copy_holder": {
            "type": "string",
            "maxLength": 300,
            "title": "Copy Holder",
            "default": ""
          },
          "copy_notes": {
            "type": "string",
            "maxLength": 10000,
            "title": "Copy Notes",
            "default": ""
          },
          "notes": {
            "type": "string",
            "maxLength": 10000,
            "title": "Notes",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ProjectCreate",
        "description": "The description of a new book; only the title is required.\n\n:ivar title: Title of the book.\n:ivar subtitle: Words of the title page that explain the title.\n:ivar parallel_titles: Titles in other languages that the title page prints beside the title.\n:ivar original_title: Title of the original work, for a translation.\n:ivar contributors: People who made the book, each with a role, in the order of the title page.\n:ivar publisher: Publisher of the book.\n:ivar printer: Printing house.\n:ivar publication_place: City of publication.\n:ivar publication_year: Year of publication as printed, which may be a range or an estimate.\n:ivar edition: Edition statement.\n:ivar censorship: Censor's permit printed in the book.\n:ivar series: Series the book belongs to.\n:ivar series_number: Number of the book in its series.\n:ivar volume: Volume or part within a multi-volume work.\n:ivar languages: ISO 639-3 codes of the languages of the text.\n:ivar orthography: Spelling system of the print.\n:ivar script: Writing system of the print.\n:ivar printed_pagination: Pagination as a catalogue states it.\n:ivar height_cm: Height of the book in centimetres.\n:ivar illustrations: Illustrations as a catalogue states them.\n:ivar binding: Binding or cover of the copy.\n:ivar identifiers: Numbers and addresses that identify the book or the copy.\n:ivar subjects: Topics of the book.\n:ivar rights: Whether the result of the work may be published.\n:ivar copy_holder: Owner of the copy that was scanned.\n:ivar copy_notes: Marks of the copy that was scanned.\n:ivar notes: Free-form notes of the owner."
      },
...
      "Problem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Problem title"
          },
          "type": {
            "type": "string",
            "title": "Problem type"
          },
          "status": {
            "type": "integer",
            "title": "Status code"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem detail"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bookreviver_session"
      }
    }
  }
}